Split AppTest into 2 files, the original AppTest.kt will contain tests that are not tied to any game, meaning they run the same regardless of game selected. Each game will have its own Test file from now on.
